Add ability to record ADSB data to a file
Reads ADSB data for a prescribed period, and writes aircraft
data to a file, with a timestamp. Allows for rounding position
data so that small changes in an aircraft's position are ignored,
which effectively compresses the data. The output is a pickled
list of timestamped position data.

The intent is to later write a player that can read this
data and play back at some faster speed.